[WIP] vdpa: Do not set iotlb map with no changes
Userspace is able to write empty iotlb batches just calling
iotlb_batch_begin/end with no map in between. Since this can be an
expensive operation for the device, it has sense to skip it.

This happens in qemu: Some versions can send empty updates if a device
is added, because although guest memory layout changes, these changes
have no meaning for the vdpa devices. Since others VMM can do the same,
let's control it here too.

This patch does not take into account that map can be the same as the
beginning, for example, if userspace adds and delete a region: It just a
simple counter.